Force conversion to sRGB for CMYK images with ICC profile
CMYK is not supported by PNG, WEBP, AVIF and HEIC Pillow encoders.

When a CMYK image is encoded, it is converted to RGB but this conversion
results in inaccurate colors because Pillow ignores the ICC profile when
performing the conversion.

As a workaround, we manually force an accurate conversion to RGB _before_
encoding the image. This results in a much more accurate representation
of the original CMYK image.